fix(web): serve the GitHub star count from our own origin
The header badge called api.github.com from the browser, which spent
the visitor's share of GitHub's unauthenticated cap (60 requests/hour
per IP) rather than ours. That holds up from a home connection and
fails from anything with a shared address — a datacentre, a corporate
proxy, carrier NAT — where the reply is a 403: a console error, and a
badge with no number.

The count now comes from /github-stars, backed by one process-wide
hourly cache. /open had a second cache of the same resource and now
shares this one, so the two surfaces no longer spend that budget
twice.

stars: null stays a valid answer — the badge drops the number, never
errors. Still unauthenticated server-side, so a busy IP degrades to
null rather than to a wrong figure.
